refactor: replace enum scope qualifiers with using enum declarations across Qt view and component files
- Add using enum declarations for QtCodeNavigator::Mode, QtLineItemBase::Route, TooltipOrigin, TokenComponentBundledEdges::Direction, MessageBookmarkButtonState::ButtonState, Graph::TrailMode, SearchMatch::SearchType, Bookmark::BookmarkFilter/Order, and CompositeView::CompositeDirection
- Remove enum class scope qualifiers throughout QtGraphEdge, QtCodeView, QtCompositeView, QtDialogView, QtGraph

+++ src/lib_gui/qt/view/QtErrorView.cpp
@@ -49,15 +49,15 @@ QtErrorView::QtErrorView(ViewLayout* viewLayout)
 	m_table->setModel(m_model);
 
 	// Setup Table Headers
-	m_model->setColumnCount(COLUMN_MAX + 1);
-	m_table->setColumnWidth(Column::ID, 40);
-	m_table->setColumnWidth(Column::TYPE, 80);
-	m_table->setColumnWidth(Column::MESSAGE, 450);
-	m_table->setColumnWidth(Column::FILE, 300);
-	m_table->setColumnWidth(Column::LINE, 50);
-	m_table->setColumnWidth(Column::TRANSLATION_UNIT, 300);
+	m_model->setColumnCount(static_cast<int>(COLUMN_MAX) + 1);
+	m_table->setColumnWidth(static_cast<int>(Column::ID), 40);
+	m_table->setColumnWidth(static_cast<int>(Column::TYPE), 80);
+	m_table->setColumnWidth(static_cast<int>(Column::MESSAGE), 450);
+	m_table->setColumnWidth(static_cast<int>(Column::FILE), 300);
+	m_table->setColumnWidth(static_cast<int>(Column::LINE), 50);
+	m_table->setColumnWidth(static_cast<int>(Column::TRANSLATION_UNIT), 300);
 
-	m_table->setColumnHidden(Column::ID, true);
+	m_table->setColumnHidden(static_cast<int>(Column::ID), true);
 
 	QStringList headers;
 	headers << QStringLiteral("ID") << QStringLiteral("Type") << QStringLiteral("Message")
@@ -68,12 +68,12 @@ QtErrorView::QtErrorView(ViewLayout* viewLayout)
 	connect(m_table, &QTableView::clicked, [=, this](const QModelIndex& index) {
 		if (index.isValid())
 		{
-			if (m_model->item(index.row(), Column::FILE) == nullptr)
+			if (m_model->item(index.row(), static_cast<int>(Column::FILE)) == nullptr)
 			{
 				return;
 			}
 
-			const Id errorId = qt_variant_cast<Id>(m_model->item(index.row(), Column::ID)->data(Qt::DisplayRole));
+			const Id errorId = qt_variant_cast<Id>(m_model->item(index.row(), static_cast<int>(Column::ID))->data(Qt::DisplayRole));
 
 			m_controllerProxy.executeAsTaskWithArgs(&ErrorController::showError, errorId);
 		}
@@ -204,7 +204,7 @@ void QtErrorView::setErrorId(Id errorId)
 {
 	m_onQtThread([=, this]() {
 		QList<QStandardItem*> items = m_model->findItems(QString::fromStdString(to_string(errorId)), 
-			Qt::MatchExactly, Column::ID);
+			Qt::MatchExactly, static_cast<int>(Column::ID));
 
 		if (items.size() == 1)
 		{
@@ -296,39 +296,39 @@ void QtErrorView::addErrorToTable(const ErrorInfo& error)
 
 	QStandardItem* item = new QStandardItem();
 	item->setData(QVariant::fromValue(error.id), Qt::DisplayRole);
-	m_model->setItem(rowNumber, Column::ID, item);
+	m_model->setItem(rowNumber, static_cast<int>(Column::ID), item);
 
 	m_model->setItem(
 		rowNumber,
-		Column::TYPE,
+		static_cast<int>(Column::TYPE),
 		new QStandardItem(error.fatal ? QStringLiteral("FATAL") : QStringLiteral("ERROR")));
 	if (error.fatal)
 	{
-		m_model->item(rowNumber, Column::TYPE)->setForeground(QBrush(Qt::red));
+		m_model->item(rowNumber, static_cast<int>(Column::TYPE))->setForeground(QBrush(Qt::red));
 	}
-	m_model->item(rowNumber, Column::TYPE)->setIcon(s_errorIcon);
+	m_model->item(rowNumber, static_cast<int>(Column::TYPE))->setIcon(s_errorIcon);
 
 	m_model->setItem(
-		rowNumber, Column::MESSAGE, new QStandardItem(QString::fromStdString(error.message)));
+		rowNumber, static_cast<int>(Column::MESSAGE), new QStandardItem(QString::fromStdString(error.message)));
 
 	m_model->setItem(
-		rowNumber, Column::FILE, new QStandardItem(QString::fromStdString(error.filePath)));
-	m_model->item(rowNumber, Column::FILE)->setToolTip(QString::fromStdString(error.filePath));
+		rowNumber, static_cast<int>(Column::FILE), new QStandardItem(QString::fromStdString(error.filePath)));
+	m_model->item(rowNumber, static_cast<int>(Column::FILE))->setToolTip(QString::fromStdString(error.filePath));
 
 	item = new QStandardItem();
 	item->setData(QVariant(qlonglong(error.lineNumber)), Qt::DisplayRole);
-	m_model->setItem(rowNumber, Column::LINE, item);
+	m_model->setItem(rowNumber, static_cast<int>(Column::LINE), item);
 
 	m_model->setItem(
 		rowNumber,
-		Column::INDEXED,
+		static_cast<int>(Column::INDEXED),
 		new QStandardItem(error.indexed ? QStringLiteral("yes") : QStringLiteral("no")));
 
 	m_model->setItem(
 		rowNumber,
-		Column::TRANSLATION_UNIT,
+		static_cast<int>(Column::TRANSLATION_UNIT),
 		new QStandardItem(QString::fromStdString(error.translationUnit)));
-	m_model->item(rowNumber, Column::TRANSLATION_UNIT)
+	m_model->item(rowNumber, static_cast<int>(Column::TRANSLATION_UNIT))
 		->setToolTip(QString::fromStdString(error.translationUnit));
 }
